refactor(clustering) refactor of clustering code to separate CP and DP logic into different modules
