+++ kong/clustering/init.lua
@@ -1,0 +1,58 @@
+local _M = {}
+
+
+local pl_file = require("pl.file")
+local pl_tablex = require("pl.tablex")
+local ssl = require("ngx.ssl")
+local openssl_x509 = require("resty.openssl.x509")
+
+
+local MT = { __index = _M, }
+
+
+function _M.new(conf)
+  assert(conf, "conf can not be nil", 2)
+
+  local self = {
+    conf = conf,
+  }
+
+  setmetatable(self, MT)
+
+  -- note: pl_file.read throws error on failure so
+  -- no need for error checking
+  local cert = pl_file.read(conf.cluster_cert)
+  self.cert = assert(ssl.parse_pem_cert(cert))
+
+  cert = openssl_x509.new(cert, "PEM")
+  self.cert_digest = cert:digest("sha256")
+
+  local key = pl_file.read(conf.cluster_cert_key)
+  self.cert_key = assert(ssl.parse_pem_priv_key(key))
+
+  self.child = require("kong.clustering." .. conf.role).new(self)
+
+  return self
+end
+
+
+function _M:handle_cp_websocket()
+  return self.child:handle_cp_websocket()
+end
+
+
+function _M:init_worker()
+  self.plugins_list = assert(kong.db.plugins:get_handlers())
+  table.sort(self.plugins_list, function(a, b)
+    return a.name:lower() < b.name:lower()
+  end)
+
+  self.plugins_list = pl_tablex.map(function(p)
+    return { name = p.name, version = p.handler.VERSION, }
+  end, self.plugins_list)
+
+  self.child:init_worker()
+end
+
+
+return _M
